requirements.txt
cairocffi==0.8.1
CairoSVG==2.1.3
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
click==6.7
coreapi==2.3.3
coreapi-cli==1.0.9
coreschema==0.0.4
cssselect2==0.2.1
defusedxml==0.5.0
Django==2.0.6
django-filter==1.1.0
django-guardian==1.4.9
django-weasyprint==0.5.2
django-wkhtmltopdf==3.1.0
djangorestframework==3.8.2
html5lib==1.0.1
httpie==0.9.9
httplib2==0.11.3
idna==2.6
itypes==1.1.0
Jinja2==2.10
Markdown==2.6.11
MarkupSafe==1.0
pdfkit==0.6.1
pdfrw==0.4
Pillow==5.1.0
pycparser==2.18
pycups==1.9.73
Pygments==2.2.0
pyPdf==1.13
PyPDF2==1.26.0
Pyphen==0.9.4
pytz==2018.4
reportlab==3.4.0
requests==2.18.4
six==1.11.0
tinycss2==0.6.1
uritemplate==3.0.0
urllib3==1.22
WeasyPrint==0.42.3
webencodings==0.5.1
xhtml2pdf==0.2.2